+++ examples/simple/program.cpp
@@ -63,6 +63,12 @@ int main()
     // add a command that prints something useful
     p.add_command("p", "print something", print_function);
 
+    // Set the ready inicator to be "# " the default is "> "
+    p.set_ready_indicator("# ");
+
+    // Set the help to be printed on run, default is on.
+    p.set_print_help(true);
+
     // Change the quit/escape key to "e", dafaults to "q"
     p.set_exit_key("e");